Finished backend auth token functionality

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 
 const asyncHandler = require('express-async-handler')
-const Goal = require ('../model/goalModel')
+const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req,res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({ user: req.user.id})
   res.status(200).json(goals)
 })
 
@@ -19,6 +20,7 @@ const setGoals = asyncHandler(async (req,res) => {
     throw new Error('Please add a text field')
   }
   const goal = await Goal.create({
+    user: req.user.id,
     text: req.body.text
   })
   res.status(200).json(goal)
@@ -37,6 +39,19 @@ const updateGoals = asyncHandler(async (req,res) => {
     res.status(400)
     throw new Error('Goal with specified ID not found')
   }
+  // find user id registered to goal
+  const user = await User.findById(req.user.id)
+  if(!user){
+    res.status(401)
+    throw new Error('Usernot found')
+  }
+
+  // verify logged in user matches goal's user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('Unauthorized')
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body,
      {new: true})
 
@@ -52,6 +67,18 @@ const deleteGoals = asyncHandler(async (req,res) => {
   if(!goal){
     res.status(400)
     throw new Error('Goal with specified ID not found')
+  }
+    // find user id registered to goal
+  const user = await User.findById(req.user.id)
+  if(!user){
+    res.status(401)
+    throw new Error('Usernot found')
+  }
+
+  // verify logged in user matches goal's user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('Unauthorized')
   }
 
   const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
